requirements.txt
anthropic>=0.49.0
streamlit>=1.43.0
pdfplumber>=0.11.0
python-docx>=1.1.0
requests>=2.32.0
python-dotenv>=1.0.0
gensim>=4.3.0
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
